+++ tests/service/optimizer/test_win_compositor.py
@@ -15,156 +15,168 @@ class GetWindowCompositorTest(IsolatedAsyncioTestCase):
     def setUp(self):
         self.expected_inxi_cmd = 'inxi -Gxx -c 0'
 
-    @patch(f'{__app_name__}.service.optimizer.win_compositor.run_async_user_process', return_value=(0, 'compositor: xpto'))
+    @patch(f'{__app_name__}.service.optimizer.win_compositor.run_async_process', return_value=(0, 0, 'compositor: xpto'))
     @patch(f'{__app_name__}.service.optimizer.win_compositor.which', return_value='/usr/bin/inxi')
-    async def test_return_none_when_inxi_returns_a_not_supported_compositor_name(self, which: Mock, run_user_process: Mock):
+    async def test_return_none_when_inxi_returns_a_not_supported_compositor_name(self, *mocks: Mock):
+        which, run_async_process = mocks
+
         env = {'abc': '123'}
         compositor = await get_window_compositor(logger=Mock(), user_id=123, user_env=env)
         which.assert_called_once_with('inxi')
-        run_user_process.assert_called_once_with(cmd=self.expected_inxi_cmd, user_id=123, user_env=env)
+        run_async_process.assert_called_once_with(cmd=self.expected_inxi_cmd, user_id=123, custom_env=env)
         self.assertIsNone(compositor)
 
-    @patch(f'{__app_name__}.service.optimizer.win_compositor.run_async_user_process', return_value=(0, """Device-1: InnoTek Systemberatung VirtualBox Graphics Adapter 
+    @patch(f'{__app_name__}.service.optimizer.win_compositor.run_async_process', return_value=(0, 0, """Device-1: InnoTek Systemberatung VirtualBox Graphics Adapter 
   driver: vboxvideo v: kernel bus-ID: 00:02.0 chip-ID: 80ee:beef 
   Display: x11 server: X.Org 1.20.11 compositor: kwin_x11 driver: 
   loaded: modesetting alternate: fbdev,vboxvideo,vesa"""))
     @patch(f'{__app_name__}.service.optimizer.win_compositor.which', return_value='/usr/bin/inxi')
-    async def test__return_kwin_when_inxi_returns_it_as_compositor(self, which: Mock, run_user_process: Mock):
+    async def test__return_kwin_when_inxi_returns_it_as_compositor(self, which: Mock, run_async_process: Mock):
         env = {'abc': '123'}
         compositor = await get_window_compositor(logger=Mock(), user_id=123, user_env=env)
         which.assert_called_with('inxi')
-        run_user_process.assert_called_once_with(cmd=self.expected_inxi_cmd, user_id=123, user_env=env)
+        run_async_process.assert_called_once_with(cmd=self.expected_inxi_cmd, user_id=123, custom_env=env)
         self.assertIsNotNone(compositor)
         self.assertIsInstance(compositor, KWinCompositor)
 
-    @patch(f'{__app_name__}.service.optimizer.win_compositor.run_async_user_process', return_value=(0, """  Device-1: InnoTek Systemberatung VirtualBox Graphics Adapter 
+    @patch(f'{__app_name__}.service.optimizer.win_compositor.run_async_process', return_value=(0, 0, """  Device-1: InnoTek Systemberatung VirtualBox Graphics Adapter 
   driver: vboxvideo v: kernel bus-ID: 00:02.0 chip-ID: 80ee:beef 
   Display: x11 server: X.Org 1.20.11 compositor: xfwm4 driver: 
   loaded: modesetting alternate: fbdev,vboxvideo,vesa"""))
     @patch(f'{__app_name__}.service.optimizer.win_compositor.which', return_value='/usr/bin/inxi')
-    async def test__return_xfwm4_when_inxi_returns_it_as_compositor(self, which: Mock, run_user_process: Mock):
+    async def test__return_xfwm4_when_inxi_returns_it_as_compositor(self, which: Mock, run_async_process: Mock):
         env = {'abc': '123'}
         compositor = await get_window_compositor(logger=Mock(), user_id=123, user_env=env)
         which.assert_called_with('inxi')
-        run_user_process.assert_called_once_with(cmd=self.expected_inxi_cmd, user_id=123, user_env=env)
+        run_async_process.assert_called_once_with(cmd=self.expected_inxi_cmd, user_id=123, custom_env=env)
         self.assertIsNotNone(compositor)
         self.assertIsInstance(compositor, Xfwm4Compositor)
 
-    @patch(f'{__app_name__}.service.optimizer.win_compositor.run_async_user_process', return_value=(0, """  Device-1: InnoTek Systemberatung VirtualBox Graphics Adapter 
+    @patch(f'{__app_name__}.service.optimizer.win_compositor.run_async_process', return_value=(0, 0, """  Device-1: InnoTek Systemberatung VirtualBox Graphics Adapter 
   driver: vboxvideo v: kernel bus-ID: 00:02.0 chip-ID: 80ee:beef 
   Display: x11 server: X.Org 1.20.11 compositor: metacity driver: 
   loaded: modesetting alternate: fbdev,vboxvideo,vesa"""))
     @patch(f'{__app_name__}.service.optimizer.win_compositor.which', return_value='/usr/bin/inxi')
-    async def test__return_marco_when_inxi_returns_metacity_as_compositor(self, which: Mock, run_user_process: Mock):
+    async def test__return_marco_when_inxi_returns_metacity_as_compositor(self, which: Mock, run_async_process: Mock):
         env = {'abc': '123'}
         compositor = await get_window_compositor(logger=Mock(), user_id=123, user_env=env)
         which.assert_called_once_with('inxi')
-        run_user_process.assert_called_once_with(cmd=self.expected_inxi_cmd, user_id=123, user_env=env)
+        run_async_process.assert_called_once_with(cmd=self.expected_inxi_cmd, user_id=123, custom_env=env)
         self.assertIsNotNone(compositor)
         self.assertIsInstance(compositor, MarcoCompositor)
 
-    @patch(f'{__app_name__}.service.optimizer.win_compositor.run_async_user_process', return_value=(0, """  Device-1: InnoTek Systemberatung VirtualBox Graphics Adapter 
+    @patch(f'{__app_name__}.service.optimizer.win_compositor.run_async_process', return_value=(0, 0, """  Device-1: InnoTek Systemberatung VirtualBox Graphics Adapter 
   driver: vboxvideo v: kernel bus-ID: 00:02.0 chip-ID: 80ee:beef 
   Display: x11 server: X.Org 1.20.11 compositor: marco driver: 
   loaded: modesetting alternate: fbdev,vboxvideo,vesa"""))
     @patch(f'{__app_name__}.service.optimizer.win_compositor.which', return_value='/usr/bin/inxi')
-    async def test__return_marco_when_inxi_returns_it_as_compositor(self, which: Mock, run_user_process: Mock):
+    async def test__return_marco_when_inxi_returns_it_as_compositor(self, which: Mock, run_async_process: Mock):
         env = {'abc': '123'}
         compositor = await get_window_compositor(logger=Mock(), user_id=123, user_env=env)
         which.assert_called_once_with('inxi')
-        run_user_process.assert_called_once_with(cmd=self.expected_inxi_cmd, user_id=123, user_env=env)
+        run_async_process.assert_called_once_with(cmd=self.expected_inxi_cmd, user_id=123, custom_env=env)
         self.assertIsNotNone(compositor)
         self.assertIsInstance(compositor, MarcoCompositor)
 
-    @patch(f'{__app_name__}.service.optimizer.win_compositor.run_async_user_process', return_value=(0, """  Device-1: InnoTek Systemberatung VirtualBox Graphics Adapter 
+    @patch(f'{__app_name__}.service.optimizer.win_compositor.run_async_process', return_value=(0, 0, """  Device-1: InnoTek Systemberatung VirtualBox Graphics Adapter 
     driver: vboxvideo v: kernel bus-ID: 00:02.0 chip-ID: 80ee:beef 
     Display: x11 server: X.Org 1.20.11 compositor: compton driver: 
     loaded: modesetting alternate: fbdev,vboxvideo,vesa"""))
     @patch(f'{__app_name__}.service.optimizer.win_compositor.which', return_value='/usr/bin/inxi')
-    async def test__return_picom_when_inxi_returns_compton_as_compositor(self, which: Mock, run_user_process: Mock):
+    async def test__return_picom_when_inxi_returns_compton_as_compositor(self, which: Mock, run_async_process: Mock):
         env = {'abc': '123'}
         compositor = await get_window_compositor(logger=Mock(), user_id=123, user_env=env)
         which.assert_called_once_with('inxi')
-        run_user_process.assert_called_once_with(cmd=self.expected_inxi_cmd, user_id=123, user_env=env)
+        run_async_process.assert_called_once_with(cmd=self.expected_inxi_cmd, user_id=123, custom_env=env)
         self.assertIsNotNone(compositor)
         self.assertIsInstance(compositor, PicomCompositor)
         self.assertEqual('Compton', compositor.get_name())
 
-    @patch(f'{__app_name__}.service.optimizer.win_compositor.run_async_user_process', return_value=(0, """  Device-1: InnoTek Systemberatung VirtualBox Graphics Adapter 
+    @patch(f'{__app_name__}.service.optimizer.win_compositor.run_async_process', return_value=(0, 0, """  Device-1: InnoTek Systemberatung VirtualBox Graphics Adapter 
     driver: vboxvideo v: kernel bus-ID: 00:02.0 chip-ID: 80ee:beef 
     Display: x11 server: X.Org 1.20.11 compositor: picom driver: 
     loaded: modesetting alternate: fbdev,vboxvideo,vesa"""))
     @patch(f'{__app_name__}.service.optimizer.win_compositor.which', return_value='/usr/bin/inxi')
-    async def test__return_picom_when_inxi_returns_it_as_compositor(self, which: Mock, run_user_process: Mock):
+    async def test__return_picom_when_inxi_returns_it_as_compositor(self, which: Mock, run_async_process: Mock):
         env = {'abc': '123'}
         compositor = await get_window_compositor(logger=Mock(), user_id=123, user_env=env)
         which.assert_called_once_with('inxi')
-        run_user_process.assert_called_once_with(cmd=self.expected_inxi_cmd, user_id=123, user_env=env)
+        run_async_process.assert_called_once_with(cmd=self.expected_inxi_cmd, user_id=123, custom_env=env)
         self.assertIsNotNone(compositor)
         self.assertIsInstance(compositor, PicomCompositor)
         self.assertEqual('Picom', compositor.get_name())
 
-    @patch(f'{__app_name__}.service.optimizer.win_compositor.run_async_user_process', return_value=(0, """  Device-1: InnoTek Systemberatung VirtualBox Graphics Adapter 
+    @patch(f'{__app_name__}.service.optimizer.win_compositor.run_async_process', return_value=(0, 0, """  Device-1: InnoTek Systemberatung VirtualBox Graphics Adapter 
     driver: vboxvideo v: kernel bus-ID: 00:02.0 chip-ID: 80ee:beef 
     Display: x11 server: X.Org 1.20.11 compositor: compiz driver: 
     loaded: modesetting alternate: fbdev,vboxvideo,vesa"""))
     @patch(f'{__app_name__}.service.optimizer.win_compositor.which', return_value='/usr/bin/inxi')
-    async def test__return_compiz_when_inxi_returns_it_as_compositor(self, which: Mock, run_user_process: Mock):
+    async def test__return_compiz_when_inxi_returns_it_as_compositor(self, which: Mock, run_async_process: Mock):
         env = {'abc': '123'}
         compositor = await get_window_compositor(logger=Mock(), user_id=123, user_env=env)
         which.assert_called_once_with('inxi')
-        run_user_process.assert_called_once_with(cmd=self.expected_inxi_cmd, user_id=123, user_env=env)
+        run_async_process.assert_called_once_with(cmd=self.expected_inxi_cmd, user_id=123, custom_env=env)
         self.assertIsNotNone(compositor)
         self.assertIsInstance(compositor, CompizCompositor)
         self.assertEqual('Compiz', compositor.get_name())
 
-    @patch(f'{__app_name__}.service.optimizer.win_compositor.run_async_user_process', return_value=(0, "Device: xpto"))
+    @patch(f'{__app_name__}.service.optimizer.win_compositor.run_async_process', return_value=(0, 0, "Device: xpto"))
     @patch(f'{__app_name__}.service.optimizer.win_compositor.which', return_value='/usr/bin/inxi')
-    async def test__return_kwin_when_inxi_does_not_return_it_but_desktop_env_is_kde(self, which: Mock, run_user_process: Mock):
+    async def test__return_kwin_when_inxi_does_not_return_it_but_desktop_env_is_kde(self, *mocks: Mock):
+        which, run_async_process = mocks
+
         env = {'XDG_CURRENT_DESKTOP': 'KDE'}
         compositor = await get_window_compositor(logger=Mock(), user_id=123, user_env=env)
         which.assert_called_with('inxi')
-        run_user_process.assert_called_once_with(cmd=self.expected_inxi_cmd, user_id=123, user_env=env)
+        run_async_process.assert_called_once_with(cmd=self.expected_inxi_cmd, user_id=123, custom_env=env)
         self.assertIsNotNone(compositor)
         self.assertIsInstance(compositor, KWinCompositor)
-        run_user_process.assert_called_once_with(cmd=self.expected_inxi_cmd, user_id=123, user_env=env)
+        run_async_process.assert_called_once_with(cmd=self.expected_inxi_cmd, user_id=123, custom_env=env)
 
-    @patch(f'{__app_name__}.service.optimizer.win_compositor.run_async_user_process', return_value=(0, "Device: xpto"))
+    @patch(f'{__app_name__}.service.optimizer.win_compositor.run_async_process', return_value=(0, 0, "Device: xpto"))
     @patch(f'{__app_name__}.service.optimizer.win_compositor.which', return_value='/usr/bin/inxi')
-    async def test__return_marco_when_inxi_does_not_return_it_but_desktop_env_is_mate(self, which: Mock, run_user_process: Mock):
+    async def test__return_marco_when_inxi_does_not_return_it_but_desktop_env_is_mate(self, *mocks: Mock):
+        which, run_async_process = mocks
+
         env = {'XDG_CURRENT_DESKTOP': 'Mate'}
         compositor = await get_window_compositor(logger=Mock(), user_id=123, user_env=env)
         which.assert_called_with('inxi')
-        run_user_process.assert_called_once_with(cmd=self.expected_inxi_cmd, user_id=123, user_env=env)
+        run_async_process.assert_called_once_with(cmd=self.expected_inxi_cmd, user_id=123, custom_env=env)
         self.assertIsNotNone(compositor)
         self.assertIsInstance(compositor, MarcoCompositor)
 
-    @patch(f'{__app_name__}.service.optimizer.win_compositor.run_async_user_process', return_value=(0, "Device: xpto"))
+    @patch(f'{__app_name__}.service.optimizer.win_compositor.run_async_process', return_value=(0, 0, "Device: xpto"))
     @patch(f'{__app_name__}.service.optimizer.win_compositor.which', return_value='/usr/bin/inxi')
-    async def test__return_xfwm4_when_inxi_does_not_return_it_but_desktop_env_is_xfce(self, which: Mock, run_user_process: Mock):
+    async def test__return_xfwm4_when_inxi_does_not_return_it_but_desktop_env_is_xfce(self, *mocks):
+        which, run_async_process = mocks
+
         env = {'XDG_CURRENT_DESKTOP': 'XFCE'}
         compositor = await get_window_compositor(logger=Mock(), user_id=123, user_env=env)
         which.assert_called_with('inxi')
-        run_user_process.assert_called_once_with(cmd=self.expected_inxi_cmd, user_id=123, user_env=env)
+        run_async_process.assert_called_once_with(cmd=self.expected_inxi_cmd, user_id=123, custom_env=env)
         self.assertIsNotNone(compositor)
         self.assertIsInstance(compositor, Xfwm4Compositor)
 
-    @patch(f'{__app_name__}.service.optimizer.win_compositor.run_async_user_process', return_value=(0, "Device: xpto"))
+    @patch(f'{__app_name__}.service.optimizer.win_compositor.run_async_process', return_value=(0, 0, "Device: xpto"))
     @patch(f'{__app_name__}.service.optimizer.win_compositor.which', return_value='/usr/bin/inxi')
-    async def test__return_none_when_inxi_does_not_return_it_and_desktop_env_is_not_supported(self, which: Mock, run_user_process: Mock):
+    async def test__return_none_when_inxi_does_not_return_it_and_desktop_env_is_not_supported(self, *mocks: Mock):
+        which, run_async_process = mocks
+
         env = {'XDG_CURRENT_DESKTOP': 'Gnome'}
         compositor = await get_window_compositor(logger=Mock(), user_id=123, user_env=env)
         which.assert_called_with('inxi')
-        run_user_process.assert_called_once_with(cmd=self.expected_inxi_cmd, user_id=123, user_env=env)
+        run_async_process.assert_called_once_with(cmd=self.expected_inxi_cmd, user_id=123, custom_env=env)
         self.assertIsNone(compositor)
 
-    @patch(f'{__app_name__}.service.optimizer.win_compositor.run_async_user_process', return_value=(0, "Device: xpto"))
+    @patch(f'{__app_name__}.service.optimizer.win_compositor.run_async_process', return_value=(0, 0, "Device: xpto"))
     @patch(f'{__app_name__}.service.optimizer.win_compositor.which', return_value='/usr/bin/inxi')
-    async def test__return_none_when_inxi_does_not_return_it_and_desktop_env_is_not_available(self, which: Mock, run_user_process: Mock):
+    async def test__return_none_when_inxi_does_not_return_it_and_desktop_env_is_not_available(self, *mocks: Mock):
+        which, run_async_process = mocks
+
         env = {'abc': '123'}
         compositor = await get_window_compositor(logger=Mock(), user_id=123, user_env=env)
         which.assert_called_with('inxi')
-        run_user_process.assert_called_once_with(cmd=self.expected_inxi_cmd, user_id=123, user_env=env)
+        run_async_process.assert_called_once_with(cmd=self.expected_inxi_cmd, user_id=123, custom_env=env)
         self.assertIsNone(compositor)
 
 
@@ -411,7 +423,9 @@ class WindowCompositorNoCLITest(IsolatedAsyncioTestCase):
         async_syscall.assert_called_once_with(f"kill -9 {context.get('pid')}")
 
     @patch(f'{__app_name__}.service.optimizer.win_compositor.system.async_syscall', return_value=(1, 'error'))
-    async def test_disable__false_when_the_compositor_process_is_alive_but_could_not_be_killed(self, async_syscall: Mock):
+    async def test_disable__false_when_the_compositor_process_is_alive_but_could_not_be_killed(self, *mocks: Mock):
+        async_syscall = mocks[0]
+
         context = {'pid': 8383737}
         self.assertFalse(await self.compositor.disable(123, {'abc': '123'}, context=context))
         async_syscall.assert_called_once_with(f"kill -9 {context.get('pid')}")
@@ -419,23 +433,23 @@ class WindowCompositorNoCLITest(IsolatedAsyncioTestCase):
     async def test_enable__false_when_enable_command_is_not_defined_on_context(self):
         self.assertFalse(await self.compositor.enable(123, {'abc': '123'}, context={}))
 
-    @patch(f'{__app_name__}.service.optimizer.win_compositor.system.run_async_user_process', return_value=(1, 'error'))
-    async def test_enable__false_when_compositor_process_could_not_be_started(self, run_async_user_process: Mock):
+    @patch(f'{__app_name__}.service.optimizer.win_compositor.system.run_async_process', return_value=(0, 1, 'error'))
+    async def test_enable__false_when_compositor_process_could_not_be_started(self, run_async_process: Mock):
         user_env = {'abc': '123'}
         context = {'cmd': 'compton -b'}
 
         self.assertFalse(await self.compositor.enable(123, user_env, context))
 
-        run_async_user_process.assert_called_once_with(cmd=context['cmd'], user_id=123, user_env=user_env)
+        run_async_process.assert_called_once_with(cmd=context['cmd'], user_id=123, custom_env=user_env)
 
-    @patch(f'{__app_name__}.service.optimizer.win_compositor.system.run_async_user_process', return_value=(0, None))
-    async def test_enable__true_when_compositor_could_be_started(self, run_async_user_process: Mock):
+    @patch(f'{__app_name__}.service.optimizer.win_compositor.system.run_async_process', return_value=(0, 0, None))
+    async def test_enable__true_when_compositor_could_be_started(self, run_async_process: Mock):
         user_env = {'abc': '123'}
         context = {'cmd': 'compton -b'}
 
         self.assertTrue(await self.compositor.enable(123, user_env, context=context))
 
-        run_async_user_process.assert_called_once_with(cmd=context['cmd'], user_id=123, user_env=user_env)
+        run_async_process.assert_called_once_with(cmd=context['cmd'], user_id=123, custom_env=user_env)
 
 
 class PicomCompositorTest(IsolatedAsyncioTestCase):