+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.3.2] 2023-01-30
+### Improvements
+- replaced some subprocess calls executed in behalf of non-root users by async calls (Python's asyncio native approach)
+
+### Fixes
+- the optimizer service (as root) not able to execute some commands in behalf of non-root users (started with Python 3.10.9)
+
 ## [1.3.1] 2022-10-22
 ### Improvements
 - Optimizing the children of the target process